Use setAttribute instead of setAttributeNode
In some versions of IE setAttributeNode throws the enigmatic "Member not
found" error with particular combinations of element and attribute. For
example calling setAttributeNode on a td with a colspan attr. Go figure :)

I’m not sure why sanitize uses setAttributeNode but setAttribute doesn’t
seem to do any harm, so here we are.